pyproject.toml
[tool.poetry]
name = "lacss"
version = "0.0.0"
description = "Tools for cell segmentation"
authors = ["Ji Yu <jyu@uchc.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
xtrain = {version = "^0.4.0", extras = ["cuda12"]}
ml-collections = "^0.1.1"
tensorflow = "^2.16.1"
typer = "^0.12.3"
matplotlib = "^3.8.4"
tqdm = "^4.65.0"
imageio = "^2.9.0"
scikit-image = ">=0.19.0"
imagecodecs = "^2023.3.16"
opencv-python = "^4.8.1.78"
cloudpickle = "^3.0.0"
numpy = "<2.0"
huggingface-hub = "^0.26.1"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.1"
pytest = "^7.4.0"
grpcio-tools = "^1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.2.3"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-gen-files = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = ["**/__pycache__/**", "**/runs/**", ".venv/"]
